perf: use float32 for Visionary expand allocation
The (N, 14) expand array was float64 (~22MB per frame for 200k points).
Switching to float32 halves this to ~11MB, reducing GC pressure.
The downstream pipeline (PointConverter, binary packing) already
standardises to float32, so this avoids a redundant precision bump.

+++ app/modules/visionary/node.py
@@ -233,7 +233,7 @@ class VisionarySensor(ModuleNode):
                 # Expand to pipeline-standard (N, 14) before transform.
                 n = compact.shape[0]
                 if compact.shape[1] == 4:
-                    points = np.zeros((n, 14), dtype=np.float64)
+                    points = np.zeros((n, 14), dtype=np.float32)
                     points[:, :3] = compact[:, :3]
                     points[:, 13] = compact[:, 3]
                 else:
